+++ src/app/(admin)/_components/search/SearchFilter.tsx
@@ -1,8 +1,8 @@
 import { cn } from "@/utils";
 import React from "react";
-import RadioGroup from "./search/RadioGroup";
-import SearchGroup from "./search/SearchGroup";
-import DateGroup from "./search/DateGroup";
+import RadioGroup from "./RadioGroup";
+import SearchGroup from "./SearchGroup";
+import DateGroup from "./DateGroup";
 
 const style = {
   buttonStyle:
