+++ src/app/(admin)/dashBoard/inquiries/page.tsx
@@ -2,8 +2,8 @@
 
 import React from "react";
 import OverviewStatBox from "../../_components/OverviewStatBox";
-import SearchFilter from "../../_components/SearchFilter";
-import DetailTable from "../../_components/DetailTable";
+import DetailTable from "../../_components/table/DetailTable";
+import SearchFilter from "../../_components/search/SearchFilter";
 
 const Page = () => {
   return (
